Refactor runtime loop structure
