+++ src/heart/programs/configurations/heart_rate.py
@@ -1,5 +1,5 @@
-from heart.environment import GameLoop
 from heart.renderers.max_bpm_screen import MaxBpmScreen
+from heart.runtime.game_loop import GameLoop
 
 
 def configure(loop: GameLoop) -> None:
