+++ src/heart/programs/configurations/l_system.py
@@ -1,5 +1,5 @@
-from heart.environment import GameLoop
 from heart.renderers.l_system import LSystem, LSystemStateProvider
+from heart.runtime.game_loop import GameLoop
 
 
 def configure(loop: GameLoop) -> None:
