+++ src/heart/programs/configurations/life.py
@@ -1,5 +1,5 @@
-from heart.environment import GameLoop
 from heart.renderers.life.renderer import Life
+from heart.runtime.game_loop import GameLoop
 
 
 def configure(loop: GameLoop) -> None:
