+++ src/heart/programs/configurations/mandlebrot.py
@@ -1,5 +1,5 @@
-from heart.environment import GameLoop
 from heart.renderers.mandelbrot.scene import MandelbrotMode
+from heart.runtime.game_loop import GameLoop
 
 
 def configure(loop: GameLoop) -> None:
