+++ src/heart/programs/configurations/pacman.py
@@ -1,8 +1,8 @@
 from heart.display.color import Color
-from heart.environment import GameLoop
 from heart.renderers.pacman import PacmanGhostRenderer
 from heart.renderers.pixels import Border
 from heart.renderers.random_pixel import RandomPixel
+from heart.runtime.game_loop import GameLoop
 
 
 def configure(loop: GameLoop) -> None:
