+++ src/heart/programs/configurations/porthole_window.py
@@ -1,8 +1,8 @@
 """Configuration for the brass porthole window renderer."""
 
-from heart.environment import GameLoop
 from heart.renderers.porthole_window import (PortholeWindowRenderer,
                                              PortholeWindowStateProvider)
+from heart.runtime.game_loop import GameLoop
 
 
 def configure(loop: GameLoop) -> None:
