+++ src/heart/programs/configurations/water.py
@@ -1,5 +1,5 @@
-from heart.environment import GameLoop
 from heart.renderers.water_cube.renderer import WaterCube
+from heart.runtime.game_loop import GameLoop
 
 
 def configure(loop: GameLoop) -> None:
