+++ src/heart/programs/registry.py
@@ -3,7 +3,7 @@ import os
 from functools import cached_property
 from typing import Callable
 
-from heart.environment import GameLoop
+from heart.runtime.game_loop import GameLoop
 
 
 class ConfigurationRegistry:
