+++ tests/navigation/test_navigation.py
@@ -1,8 +1,8 @@
 import pytest
 
 from heart.display.color import Color
-from heart.environment import GameLoop
 from heart.renderers.color import RenderColor
+from heart.runtime.game_loop import GameLoop
 
 
 class TestNavigationNavigation:
